# app/api/ai/strategy/route.ts
/**
 * Strategy API Route
 *
 * Endpoint for generating brand strategy using AI providers.
 *
 * Flow:
 * 1. Authenticate user with Supabase
 * 2. Validate project ownership
 * 3. Check budget limits (project, user, daily)
 * 4. Load research data as context
 * 5. Generate strategy using LLMProviderManager
 * 6. Save output to database
 * 7. Log usage and costs
 * 8. Return response with metadata
 *
 * @route POST /api/ai/strategy
 *
 * @example
 * ```typescript
 * const response = await fetch('/api/ai/strategy', {
 *   method: 'POST',
 *   headers: {
 *     'Content-Type': 'application/json',
 *     'Authorization': `Bearer ${token}`,
 *   },
 *   body: JSON.stringify({
 *     projectId: '123',
 *     researchId: '456',
 *     brief: 'Create positioning strategy...',
 *   }),
 * });
 * ```
 */

import { NextRequest, NextResponse } from 'next/server';
import { cookies } from 'next/headers';
import { requireProjectOwnership, handleAuthError } from '@/lib/auth/middleware';
import { checkBudgets, logLLMUsage, saveLLMOutput } from '@/lib/auth/helpers';
import { createClient } from '@/lib/supabase/server';

/**
 * Request body interface
 */
interface StrategyRequest {
  projectId: string;
  researchId?: string; // Optional: use existing research
  researchData?: string; // Optional: provide research directly
  brief: string;
  options?: {
    model?: string;
    provider?: string;
    temperature?: number;
    maxTokens?: number;
  };
}

/**
 * Response interface
 */
interface StrategyResponse {
  success: boolean;
  data?: {
    strategy: string;
    metadata: {
      model: string;
      provider: string;
      tokensUsed: number;
      cost: number;
      latencyMs: number;
      cached: boolean;
    };
  };
  error?: {
    code: string;
    message: string;
  };
}

/**
 * POST /api/ai/strategy
 *
 * Generate brand strategy from brief and research.
 */
export async function POST(request: NextRequest) {
  try {
    // 1. Parse and validate request body
    const body: StrategyRequest = await request.json();

    if (!body.projectId || !body.brief) {
      return NextResponse.json(
        {
          success: false,
          error: {
            code: 'INVALID_REQUEST',
            message: 'projectId and brief are required',
          },
        },
        { status: 400 }
      );
    }

    // 2. Authenticate and verify project ownership
    const { user, project, error: authError } = await requireProjectOwnership(body.projectId);
    if (authError) return authError;

    // 3. Load research data if researchId provided
    let researchContext = body.researchData || '';

    if (body.researchId && !researchContext) {
      const supabase = createClient(cookies());
      const { data: research, error: fetchError } = await supabase
        .from('outputs')
        .select('content')
        .eq('id', body.researchId)
        .eq('project_id', body.projectId)
        .eq('section', 'research')
        .single();

      if (!fetchError && research) {
        researchContext = (research as { content: string }).content;
      }
    }

    // 4. Check budget limits (strategy tasks use more tokens)
    const budgetCheck = await checkBudgets(body.projectId, user.id);
    if (!budgetCheck.allowed) {
      return NextResponse.json(
        {
          success: false,
          error: {
            code: 'BUDGET_EXCEEDED',
            message: budgetCheck.message || 'Budget exceeded',
          },
        },
        { status: 429 }
      );
    }

    // TODO: 6. Get LLM Provider Manager
    // import { getLLMProviderManager } from '@/lib/ai/manager';
    // const manager = getLLMProviderManager();

    // TODO: 7. Load strategy prompt template
    // import { createStrategyPrompt } from '@/lib/ai/prompts/strategy';
    // const prompt = createStrategyPrompt({
    //   brief: body.brief,
    //   research: researchContext,
    //   industry: project.data.industry,
    //   targetAudience: project.data.target_audience,
    // });

    // TODO: 8. Generate strategy
    // Strategy tasks use premium models (Claude Opus or Grok Beta)
    // const result = await manager.generateCompletion({
    //   prompt,
    //   systemPrompt: 'You are a senior brand strategist...',
    //   taskType: 'strategy',
    //   model: body.options?.model,
    //   provider: body.options?.provider,
    //   temperature: body.options?.temperature ?? 0.8,
    //   maxTokens: body.options?.maxTokens ?? 8192,
    //   metadata: {
    //     projectId: body.projectId,
    //     userId: session.user.id,
    //     taskType: 'strategy',
    //   },
    // });

    // TODO: 9. Save strategy to database
    // await supabase.from('strategy_outputs').insert({
    //   project_id: body.projectId,
    //   research_id: body.researchId,
    //   content: result.content,
    //   model: result.model,
    //   provider: result.providerId,
    //   tokens_used: result.usage.totalTokens,
    //   cost: result.cost,
    //   latency_ms: result.latencyMs,
    //   created_by: session.user.id,
    // });

    // TODO: 10. Log usage for analytics
    // await supabase.from('llm_usage_logs').insert({
    //   user_id: session.user.id,
    //   project_id: body.projectId,
    //   provider: result.providerId,
    //   model: result.model,
    //   task_type: 'strategy',
    //   prompt_tokens: result.usage.promptTokens,
    //   completion_tokens: result.usage.completionTokens,
    //   total_tokens: result.usage.totalTokens,
    //   cost: result.cost,
    //   latency_ms: result.latencyMs,
    //   cached: result.cached,
    //   success: true,
    // });

    // TODO: 11. Return success response
    return NextResponse.json({
      success: true,
      data: {
        strategy: 'TODO: Implement strategy generation',
        metadata: {
          model: 'claude-3-5-sonnet-20241022',
          provider: 'claude',
          tokensUsed: 0,
          cost: 0,
          latencyMs: 0,
          cached: false,
        },
      },
    });

  } catch (error) {
    console.error('Strategy generation error:', error);
    return handleAuthError(error);
  }
}

/**
 * GET /api/ai/strategy
 *
 * Get previous strategies for a project.
 */
export async function GET(request: NextRequest): Promise<NextResponse> {
  try {
    const { searchParams } = new URL(request.url);
    const projectId = searchParams.get('projectId');

    if (!projectId) {
      return NextResponse.json(
        { success: false, error: { code: 'INVALID_REQUEST', message: 'projectId is required' } },
        { status: 400 }
      );
    }

    // Authenticate and verify project ownership
    const { user, error: authError } = await requireProjectOwnership(projectId);
    if (authError) return authError;

    // TODO: Fetch strategy outputs from database
    // const supabase = createClient(cookies());
    // const { data: strategies } = await supabase
    //   .from('outputs')
    //   .select('*')
    //   .eq('project_id', projectId)
    //   .eq('section', 'strategy')
    //   .order('created_at', { ascending: false })
    //   .limit(10);

    return NextResponse.json({
      success: true,
      data: {
        strategies: [],
      },
    });

  } catch (error) {
    console.error('Strategy fetch error:', error);
    return handleAuthError(error);
  }
}